- formal changes - fixed err message format string in Sensofar
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@8928 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ modules/file/sensofar.c
@@ -170,8 +170,8 @@ sensofar_load(const gchar *filename,
    }
    if (size < sizeof(tDataDesc) + 12) {
       g_set_error(error, GWY_MODULE_FILE_ERROR, GWY_MODULE_FILE_ERROR_DATA,
-                  _("File too small (less than %d bytes). Header is not complete."), 
-                  sizeof(tDataDesc)+12);
+                  _("File too small (less than %lu bytes). Header is not complete."), 
+                  (gulong)(sizeof(tDataDesc)+12));
       return NULL;
    }
    // A little dirty, set pointer to data description structure to start of buffer
